requirements.txt
websocket-client
tornado
configparser
